requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==23.2
djangorestframework==3.15.2
drf-yasg==1.21.5
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
